pyproject.toml
[tool.poetry]
name = "web3-bot"
version = "0.1.0"
description = ""
authors = ["Naman Tamrakar <namantam1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "v20.0a4"
gspread = "^5.5.0"
oauth2client = "^4.1.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
